Sort of got server side rendering to work
