+++ app/api/mock_api.ts
@@ -1,11 +1,10 @@
 import fs from 'fs';
 import path from 'path';
 
-const mockDataFolderPath = path.join(__dirname, 'components/mock_data');
-
 export function fetchAirbudsFromServer(): Promise<any> {
   return new Promise((resolve, reject) => {
-    const filePath = path.join(mockDataFolderPath, 'airbuds_data.json');
+    const filePath =
+      '/Users/vivekgarg/Documents/CS35L/Project/cs35l-big-backs/components/mock_data/airbuds_data.json';
     fs.readFile(filePath, 'utf8', (err, data) => {
       if (err) {
         reject(err);
@@ -18,12 +17,15 @@ export function fetchAirbudsFromServer(): Promise<any> {
 
 export function fetchArtistsFromServer(): Promise<any> {
   return new Promise((resolve, reject) => {
-    const filePath = path.join(mockDataFolderPath, 'artist_data.json');
+    const filePath =
+      '/Users/vivekgarg/Documents/CS35L/Project/cs35l-big-backs/components/mock_data/artist_data.json';
     fs.readFile(filePath, 'utf8', (err, data) => {
       if (err) {
         reject(err);
       } else {
-        resolve(JSON.parse(data));
+        setTimeout(() => {
+          resolve(JSON.parse(data));
+        }, 5000);
       }
     });
   });
@@ -31,7 +33,8 @@ export function fetchArtistsFromServer(): Promise<any> {
 
 export function fetchPlaylistsFromServer(): Promise<any> {
   return new Promise((resolve, reject) => {
-    const filePath = path.join(mockDataFolderPath, 'playlist_data.json');
+    const filePath =
+      '/Users/vivekgarg/Documents/CS35L/Project/cs35l-big-backs/components/mock_data/playlist_data.json';
     fs.readFile(filePath, 'utf8', (err, data) => {
       if (err) {
         reject(err);
@@ -44,7 +47,8 @@ export function fetchPlaylistsFromServer(): Promise<any> {
 
 export function fetchSongsFromServer(): Promise<any> {
   return new Promise((resolve, reject) => {
-    const filePath = path.join(mockDataFolderPath, 'song_data.json');
+    const filePath =
+      '/Users/vivekgarg/Documents/CS35L/Project/cs35l-big-backs/components/mock_data/song_data.json';
     fs.readFile(filePath, 'utf8', (err, data) => {
       if (err) {
         reject(err);
